// src/hooks/api/carRentalRequests/carRentalRequests.generated.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** Scalar for receiving email from client */
  Email: any;
};
// ...
export type BundleSubscriptionResellerDiscountType = {
  __typename?: 'BundleSubscriptionResellerDiscountType';
  bundleType: BundleTypeType;
  carMakerName?: Maybe<Scalars['String']>;
  carModelName?: Maybe<Scalars['String']>;
  createdAt: Scalars['Float'];
  createdBy?: Maybe<User>;
  estimatedCarDeliveryDate?: Maybe<Scalars['String']>;
  grantee: BundleSubscriptionResellerDiscountGranteeType;
  id: Scalars['String'];
  isCarPartOfFleet?: Maybe<Scalars['Boolean']>;
  notes: Array<BundleInviteInternalNoteType>;
  resellerBusiness: BusinessType;
  resellerPrice?: Maybe<Scalars['Int']>;
  status: Scalars['String'];
};
// ...
export type BusinessType = {
  __typename?: 'BusinessType';
  address: Scalars['String'];
  apiKey: Scalars['String'];
  brandName: Scalars['String'];
  companyNumber: Scalars['String'];
  displayName: Scalars['String'];
  emailAddress: Scalars['String'];
  id: Scalars['String'];
  isFleetInsuranceInPlace: Scalars['Boolean'];
  locationName: Scalars['String'];
  logoImage?: Maybe<ImageType>;
  name: Scalars['String'];
  phoneNumber: Scalars['String'];
  reviews: Array<ReviewType>;
  reviewsAverageScore?: Maybe<Scalars['Float']>;
  reviewsCount: Scalars['Int'];
  tradingHours: Scalars['String'];
  type: BusinessTypeEnum;
  users: Array<User>;
};

export enum BusinessTypeEnum {
  BundlePartner = 'BundlePartner',
  Insurer = 'Insurer',
  Reseller = 'Reseller'
}
// ...
export enum BusinessUserRoleEnum {
  Operator = 'Operator',
  Owner = 'Owner'
}
// ...
export type CallRosterRequestType = {
  __typename?: 'CallRosterRequestType';
  createdAt: Scalars['Float'];
  id: Scalars['String'];
  requestStatus: CallRequestStatusEnum;
  requestedDate: Scalars['Float'];
  requestedTime: CallRequestRequestedTimeTypeEnum;
  /** returns the reseller name of the subscription the user have for this call roster request */
  resellerName?: Maybe<Scalars['String']>;
  /** returns the user attached to the call roster request */
  user: User;
};

export type CallRosterRequestUpdateInput = {
  id: Scalars['String'];
  requestedDate: Scalars['Float'];
  requestedTime: CallRequestRequestedTimeTypeEnum;
};

export type CarAvailabilityPeriodComputedType = {
  __typename?: 'CarAvailabilityPeriodComputedType';
  timeEnd: Scalars['Float'];
  timeStart: Scalars['Float'];
};

export type CarAvailabilityPeriodType = {
  __typename?: 'CarAvailabilityPeriodType';
  allDayDate?: Maybe<Scalars['String']>;
  chainId?: Maybe<Scalars['String']>;
  id: Scalars['String'];
  timeEnd?: Maybe<Scalars['Float']>;
  timeStart?: Maybe<Scalars['Float']>;
};

export enum CarDetailsApproveStatusEnum {
  Approved = 'Approved',
  Pending = 'Pending',
  Rejected = 'Rejected'
}

export type CarDetailsType = {
  __typename?: 'CarDetailsType';
  approveStatus?: Maybe<CarDetailsApproveStatusEnum>;
  color?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  doors?: Maybe<Scalars['Int']>;
  id: Scalars['String'];
  images: Array<ImageType>;
  mainImageId?: Maybe<Scalars['String']>;
  maker?: Maybe<CarMakerType>;
  mileage?: Maybe<Scalars['Int']>;
  model?: Maybe<CarModelType>;
  motor?: Maybe<CarMotorEnum>;
  range?: Maybe<CarRangeEnum>;
  registration: Scalars['String'];
  rejectionReason?: Maybe<Scalars['String']>;
  seats?: Maybe<Scalars['Int']>;
  transmission?: Maybe<CarTransmissionTypeEnum>;
  type?: Maybe<CarTypeEnum>;
  value?: Maybe<CarValueEnum>;
  vin?: Maybe<Scalars['String']>;
  year?: Maybe<Scalars['Int']>;
};
// ...
export type CarMakerType = {
  __typename?: 'CarMakerType';
  id: Scalars['String'];
  models: Array<CarModelType>;
  name: Scalars['String'];
  visible: Scalars['Boolean'];
};

export type CarModelType = {
  __typename?: 'CarModelType';
  id: Scalars['String'];
  name: Scalars['String'];
  visible: Scalars['Boolean'];
};

export enum CarMotorEnum {
  Electric = 'Electric',
  Hybrid = 'Hybrid'
}
// ...
export enum CarRentalRequestPaymentStatusEnum {
  Completed = 'Completed',
  Failed = 'Failed',
  InProgress = 'InProgress',
  None = 'None',
  RequiresAction = 'RequiresAction'
}

export type CarRentalRequestPricesType = {
  __typename?: 'CarRentalRequestPricesType';
  additionalFee: Scalars['Int'];
  rentalFeeOwner: Scalars['Int'];
  rentalFeeSum: Scalars['Int'];
  rentalFeeZoom: Scalars['Int'];
};

export enum CarRentalRequestStatusEnum {
  Accepted = 'Accepted',
  Cancelled = 'Cancelled',
  Finished = 'Finished',
  Incidented = 'Incidented',
  Rejected = 'Rejected',
  WaitingForAcceptance = 'WaitingForAcceptance'
}

export type CarRentalRequestType = {
  __typename?: 'CarRentalRequestType';
  acceptanceNote?: Maybe<Scalars['String']>;
  additionalFee: Scalars['Int'];
  cancellationNote?: Maybe<Scalars['String']>;
  car: CarType;
  createdAt: Scalars['Float'];
  id: Scalars['String'];
  incidentDescription?: Maybe<Scalars['String']>;
  internalNote?: Maybe<Scalars['String']>;
  paymentMethod?: Maybe<PaymentMethodType>;
  paymentStatus: CarRentalRequestPaymentStatusEnum;
  prices: CarRentalRequestPricesType;
  rejectionNote?: Maybe<Scalars['String']>;
  rentalFeeOwner: Scalars['Int'];
  rentalFeeSum: Scalars['Int'];
  rentalFeeZoom: Scalars['Int'];
  status: CarRentalRequestStatusEnum;
  stripePaymentIntentClientSecret?: Maybe<Scalars['String']>;
  stripePaymentIntentId?: Maybe<Scalars['String']>;
  stripePaymentIntentUrl?: Maybe<Scalars['String']>;
  stripeSetupIntentId?: Maybe<Scalars['String']>;
  stripeSetupIntentUrl?: Maybe<Scalars['String']>;
  timeEnd: Scalars['Float'];
  timeStart: Scalars['Float'];
  user: User;
};

export type CarRentalRequestsQueryParamsInput = {
  limit?: InputMaybe<Scalars['Int']>;
  offset?: InputMaybe<Scalars['Int']>;
  ownerBusinessId?: InputMaybe<Scalars['String']>;
  ownerUserId?: InputMaybe<Scalars['String']>;
  renterUserId?: InputMaybe<Scalars['String']>;
  requestingAsRenter?: InputMaybe<Scalars['Boolean']>;
  status?: InputMaybe<CarRentalRequestStatusEnum>;
};

export type CarRentalRequestsQueryResultType = {
  __typename?: 'CarRentalRequestsQueryResultType';
  overallCount: Scalars['Int'];
  rentalRequests: Array<CarRentalRequestType>;
};
// ...
export type CarType = {
  __typename?: 'CarType';
  addressPrivate?: Maybe<Scalars['String']>;
  addressPublic?: Maybe<Scalars['String']>;
  availabilityPeriods: Array<CarAvailabilityPeriodType>;
  availabilityPeriodsComputed: Array<CarAvailabilityPeriodComputedType>;
  bookedPeriods: Array<CarAvailabilityPeriodType>;
  business?: Maybe<BusinessType>;
  createdAt: Scalars['Float'];
  details: CarDetailsType;
  detailsRequested?: Maybe<CarDetailsType>;
  fakeLocation?: Maybe<CarFakeLocationType>;
  features: Array<CarFeatureType>;
  flags: CarFlagsType;
  id: Scalars['String'];
  isAvailableToBuy: Scalars['Boolean'];
  isDraft: Scalars['Boolean'];
  pricePerDay?: Maybe<Scalars['Int']>;
  pricePerHour?: Maybe<Scalars['Int']>;
  pricePerWeek?: Maybe<Scalars['Int']>;
  realLocation?: Maybe<CarRealLocationType>;
  reviews: Array<ReviewType>;
  reviewsAverageScore?: Maybe<Scalars['Float']>;
  reviewsCount: Scalars['Int'];
  salesInfo?: Maybe<CarSalesInfoType>;
  user?: Maybe<User>;
  visible: Scalars['Boolean'];
};
// ...
export type CommunicationPreferencesType = {
  __typename?: 'CommunicationPreferencesType';
  email: Scalars['String'];
  emailAllowed: Scalars['Boolean'];
  id: Scalars['String'];
  phoneAllowed: Scalars['Boolean'];
  smsAllowed: Scalars['Boolean'];
};
// ...
export type ImageType = {
  __typename?: 'ImageType';
  id: Scalars['String'];
  url: Scalars['String'];
};
// ...
export type PaymentMethodType = {
  __typename?: 'PaymentMethodType';
  brand?: Maybe<Scalars['String']>;
  id: Scalars['String'];
  last4: Scalars['String'];
};
// ...
export type Query = {
// ...
  carRentalRequests: CarRentalRequestsQueryResultType;
// ...
  user?: Maybe<User>;
  /** Returns list of users queried by given params */
  users: Array<User>;
  /** Returns count of users present in database */
  usersCount: Scalars['Int'];
  /** Returns list of users queried by given params and overallCount */
  usersV2: UsersV2QueryResultType;
};
// ...
export type ReviewType = {
  __typename?: 'ReviewType';
  business?: Maybe<BusinessType>;
  id: Scalars['String'];
  note: Scalars['String'];
  resourceId: Scalars['String'];
  reviewType: ReviewTypeEnum;
  score: Scalars['Int'];
  user?: Maybe<User>;
};
// ...
export type TagType = {
  __typename?: 'TagType';
  id: Scalars['String'];
  name: Scalars['String'];
};
// ...
export type User = {
  __typename?: 'User';
  bannedAt?: Maybe<Scalars['Float']>;
  business?: Maybe<BusinessType>;
  businessUserRole?: Maybe<BusinessUserRoleEnum>;
  callRequest?: Maybe<CallRosterRequestType>;
  communicationPreferences?: Maybe<CommunicationPreferencesType>;
  createdAt: Scalars['Float'];
  details: UserDetailsType;
  detailsRequested?: Maybe<UserDetailsType>;
  email: Scalars['String'];
  id: Scalars['String'];
  internalDetails: UserInternalDetailsType;
  invites: Array<BundleSubscriptionResellerDiscountType>;
  isAdmin: Scalars['Boolean'];
  isBusinessInvitationPopupClosed: Scalars['Boolean'];
  isOwnerProfileSetupFlowInitiated: Scalars['Boolean'];
  /** @deprecated please use communicationPreferences instead */
  marketingPermissionsProvided: Scalars['Boolean'];
  paymentMethods: Array<PaymentMethodType>;
  reviews: Array<ReviewType>;
  reviewsAverageScore?: Maybe<Scalars['Float']>;
  reviewsCount: Scalars['Int'];
  stripeConnectedAccountId?: Maybe<Scalars['String']>;
  stripeConnectedAccountSetupRequired?: Maybe<Scalars['Boolean']>;
  stripeConnectedAccountUrl?: Maybe<Scalars['String']>;
  stripeCustomerId?: Maybe<Scalars['String']>;
  stripeCustomerUrl?: Maybe<Scalars['String']>;
  subscriptions: Array<BundleSubscriptionType>;
  tags: Array<TagType>;
  validAsCarOwner: Scalars['Boolean'];
  validAsCarRenter: Scalars['Boolean'];
};


export type UserStripeCustomerIdArgs = {
  createIfNotExist?: InputMaybe<Scalars['Boolean']>;
};

export type UserDeleteType = {
  __typename?: 'UserDeleteType';
  deleted: Scalars['Boolean'];
};

export enum UserDetailsApprovalStatusEnum {
  Approved = 'Approved',
  Draft = 'Draft',
  Pending = 'Pending',
  Rejected = 'Rejected'
}

export type UserDetailsType = {
  __typename?: 'UserDetailsType';
  addressLine1?: Maybe<Scalars['String']>;
  addressLine2?: Maybe<Scalars['String']>;
  approvalStatus?: Maybe<UserDetailsApprovalStatusEnum>;
  avatarImage?: Maybe<ImageType>;
  checkCode?: Maybe<Scalars['String']>;
  city?: Maybe<Scalars['String']>;
  county?: Maybe<Scalars['String']>;
  dateOfBirth?: Maybe<Scalars['String']>;
  drivingLicenseCountryOfIssue?: Maybe<Scalars['String']>;
  drivingLicenseFrontImage?: Maybe<ImageType>;
  drivingLicenseNumber?: Maybe<Scalars['String']>;
  drivingLicenseSelfieImage?: Maybe<ImageType>;
  drivingLicenseValidFrom?: Maybe<Scalars['String']>;
  drivingLicenseValidTo?: Maybe<Scalars['String']>;
  drivingRecordCheckCode?: Maybe<Scalars['String']>;
  firstName?: Maybe<Scalars['String']>;
  lastName?: Maybe<Scalars['String']>;
  passportFrontImage?: Maybe<ImageType>;
  phoneNumber?: Maybe<Scalars['String']>;
  postCode?: Maybe<Scalars['String']>;
  rejectionReason?: Maybe<Scalars['String']>;
};
// ...
export type UserInternalDetailsType = {
  __typename?: 'UserInternalDetailsType';
  isThreeWayVerified?: Maybe<Scalars['Boolean']>;
  notes: Array<UserInternalNoteType>;
  numberOfPoints?: Maybe<Scalars['String']>;
  restrictionsOnLicense?: Maybe<Scalars['String']>;
  userId: Scalars['String'];
};
// ...
export type CarRentalRequestsQuery = { __typename?: 'Query', carRentalRequests: { __typename?: 'CarRentalRequestsQueryResultType', overallCount: number, rentalRequests: Array<{ __typename?: 'CarRentalRequestType', id: string, status: Types.CarRentalRequestStatusEnum, timeStart: number, timeEnd: number, additionalFee: number, rentalFeeSum: number, rentalFeeZoom: number, rentalFeeOwner: number, acceptanceNote?: string | null, rejectionNote?: string | null, incidentDescription?: string | null, createdAt: number, paymentStatus: Types.CarRentalRequestPaymentStatusEnum, user: { __typename?: 'User', id: string, reviewsAverageScore?: number | null, reviewsCount: number, details: { __typename?: 'UserDetailsType', firstName?: string | null, lastName?: string | null, avatarImage?: { __typename?: 'ImageType', url: string, id: string } | null }, business?: { __typename?: 'BusinessType', displayName: string, id: string, logoImage?: { __typename?: 'ImageType', url: string } | null } | null }, car: { __typename?: 'CarType', id: string, business?: { __typename?: 'BusinessType', id: string, displayName: string, logoImage?: { __typename?: 'ImageType', id: string, url: string } | null } | null, user?: { __typename?: 'User', id: string, details: { __typename?: 'UserDetailsType', firstName?: string | null, lastName?: string | null, avatarImage?: { __typename?: 'ImageType', id: string, url: string } | null } } | null, details: { __typename?: 'CarDetailsType', registration: string, mainImageId?: string | null, maker?: { __typename?: 'CarMakerType', name: string } | null, model?: { __typename?: 'CarModelType', name: string } | null, images: Array<{ __typename?: 'ImageType', url: string, id: string }> } }, prices: { __typename?: 'CarRentalRequestPricesType', rentalFeeSum: number, rentalFeeOwner: number, rentalFeeZoom: number, additionalFee: number } }> } };